% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.getNbdoses}
\alias{.getNbdoses}
\title{Get the number of doses for steady state for monolix2rx}
\usage{
.getNbdoses(x)
}
\arguments{
\item{x}{monolix2rx object, its parsed project or its rxUi}
}
\value{
Number of doses for monolix2rx object
}
\description{
Get the number of doses for steady state for monolix2rx
}
\author{
Matthew L. Fidler
}
\keyword{internal}
