% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterUpdate.R
\name{.monolixGetMlxtran}
\alias{.monolixGetMlxtran}
\title{Get the best mlxtran info or return NULL}
\usage{
.monolixGetMlxtran(x)
}
\arguments{
\item{x}{item to try to extract mlxtran info}
}
\value{
mlxtran info
}
\description{
Get the best mlxtran info or return NULL
}
\author{
Matthew L. Fidler
}
\keyword{internal}
