% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCov.R
\name{mlxtranGetMutate}
\alias{mlxtranGetMutate}
\title{Get the string of the mutate statement for input dataset based on
mlxtran}
\usage{
mlxtranGetMutate(mlxtran)
}
\arguments{
\item{mlxtran}{input mlxtran file}
}
\value{
mlxtran string that can be applied to a model (by evaluating it)
}
\description{
Get the string of the mutate statement for input dataset based on
mlxtran
}
\examples{

covD <- system.file("cov", package="monolix2rx")
m <- mlxtran(file.path(covD, "phenobarbital_project.mlxtran"))
message(mlxtranGetMutate(m))
}
\author{
Matthew L. Fidler
}
