% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{monolixDataImport}
\alias{monolixDataImport}
\title{Import a dataset from monolix (based on an imported model)}
\usage{
monolixDataImport(ui, data, na.strings = c("NA", "."))
}
\arguments{
\item{ui}{A rxode2 ui model imported from monolix}

\item{data}{dataset to convert to nlmixr2 format; if the dataset is
missing, load from the mlxtran specification}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after}
    white space is stripped from the input, so \code{na.strings}
    values may need their own white space stripped in advance.}
}
\value{
Dataset appropriate for using with the rxode2 model for
simulations
}
\description{
Import a dataset from monolix (based on an imported model)
}
\author{
Matthew L. Fidler
}
