% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpointInfo.R
\name{monolixEndpoints}
\alias{monolixEndpoints}
\title{This gets the monolix endpoint information in the model}
\usage{
monolixEndpoints(mlxtran)
}
\arguments{
\item{mlxtran}{This is the parsed mlxtran or imported rxode2 model from monolix}
}
\value{
nothing, called for side effects
}
\description{
This gets the monolix endpoint information in the model
}
\author{
Matthew L. Fidler
}
