% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etaImport.R
\name{monolixEtaImport}
\alias{monolixEtaImport}
\title{Load data from a mlxtran defined dataset}
\usage{
monolixEtaImport(mlxtran, na.strings = c("NA", "."))
}
\arguments{
\item{mlxtran}{mlxtran file where data input is specified}

\item{na.strings}{a character vector of strings which are to be
    interpreted as \code{\link{NA}} values.  Blank fields are also
    considered to be missing values in logical, integer, numeric and
    complex fields.  Note that the test happens \emph{after}
    white space is stripped from the input, so \code{na.strings}
    values may need their own white space stripped in advance.}
}
\description{
Load data from a mlxtran defined dataset
}
\author{
Matthew L Fidler
}
