% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.getStiff}
\alias{.getStiff}
\title{Get if the model object is stiff}
\usage{
.getStiff(x)
}
\arguments{
\item{x}{monolix2rx object}
}
\value{
boolean indicating if the object is a stiff system (as indicated by monolix)
}
\description{
Get if the model object is stiff
}
\author{
Matthew L. Fidler
}
\keyword{internal}
