% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlxtran.R
\name{mlxtran}
\alias{mlxtran}
\title{Read and parse mlxtran lines}
\usage{
mlxtran(file, equation = FALSE)
}
\arguments{
\item{file}{mlxtran file to process}

\item{equation}{parse the equation block to rxode2 (some models cannot be translated)}
}
\value{
mlxtran object
}
\description{
Read and parse mlxtran lines
}
\author{
Matthew L. Fidler
}
