% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation.R
\name{mlxTxt}
\alias{mlxTxt}
\title{Get equation block from a Monolix model txt file}
\usage{
mlxTxt(file, retFile = FALSE)
}
\arguments{
\item{file}{string representing the model text file.  Can be
lib:fileName.txt if library setup/available}

\item{retFile}{boolean that tells \code{mlxTxt()} to return the file
name instead of error if the file does not exist}
}
\value{
parsed equation or file name
}
\description{
Get equation block from a Monolix model txt file
}
\author{
Matthew L. Fidler
}
