% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monolix2rx.R
\name{monolix2rx}
\alias{monolix2rx}
\title{Translate a monolix file to rxode2}
\usage{
monolix2rx(mlxtran)
}
\arguments{
\item{mlxtran}{file name for mlxtran to translate to rxode2}
}
\value{
rxode2 model
}
\description{
Translate a monolix file to rxode2
}
\author{
Matthew L. Fidler
}
